# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterRequest {
  email: string;
  /** @minLength 2 */
  username: string;
  /** @minLength 6 */
  password: string;
  /** Date of birth in YYYY-MM-DD format */
  dateOfBirth?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface AuthUser {
  id: number;
  email: string;
  username: string;
  dateOfBirth?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: AuthUser;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}